% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfd.R
\name{get_tfd}
\alias{get_tfd}
\title{Read file}
\usage{
get_tfd(file)
}
\arguments{
\item{file}{path to a pdf file}
}
\value{
a string vector, where every string is a page in the process.
}
\description{
Reads a penal execution process pdf into R
}
\examples{
\dontrun{
  get_pdf("data-raw/fulano.pdf")
}
}
